+++ CHANGELOG.md
@@ -3,9 +3,7 @@
 ## 1.6.0
 * Add support for Local arguments in kernel functions
 * Add full support for OpenCL 1.2 atomic operations on arrays of integers.
-
-## 1.5.1
-
+* Parent pom no longer points to a snapshot.
 
 ## 1.5.0